pyproject.toml
[project]
name = "tcc-genai-rag"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.5",
  "uvicorn==0.32.0",
  "httpx==0.27.2",
  "pydantic==2.9.2",
  "pytest==8.3.3",
  "pytest-cov==5.0.0",
  "radon==6.0.1",
  "pylint==3.3.1",
  "ruff==0.6.9",
  "bandit==1.7.10",
  "faiss-cpu==1.8.0.post1",
  "sentence-transformers==3.2.1",
  "python-dotenv==1.0.1",
  "typer==0.12.5",
  "rich==13.9.2",
  "ruyaml==0.91.0",
]

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings"


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


